Moved header/footer/logo/settingDropdown from _components/ui to _components/layout
