+++ client/app/_components/layout/header/stories/Header.stories.ts
@@ -2,7 +2,7 @@ import { Header } from '../Header'
 import type { Meta, StoryObj } from '@storybook/react'
 
 const meta = {
-  title: 'ui/Header',
+  title: 'layout/Header',
   component: Header,
   parameters: {
     layout: 'fullscreen',
