+++ client/app/_components/layout/index.ts
@@ -1,0 +1,4 @@
+export { SettingDropdown } from './SettingDropdown/SettingDropdown'
+export { Header } from './header/Header'
+export { Logo } from './logo/Logo'
+export { Footer } from './footer/Footer'
