+++ client/app/_components/layout/logo/stories/Logo.stories.tsx
@@ -3,7 +3,7 @@ import { Logo } from '../Logo'
 import type { Meta, StoryObj } from '@storybook/react'
 
 const meta: Meta<typeof Logo> = {
-  title: 'ui/Logo',
+  title: 'layout/Logo',
   component: Logo,
   parameters: {
     layout: 'centered',
