+++ client/app/_components/ui/index.ts
@@ -1,6 +1,8 @@
 export { Button, type ButtonProps } from './button/Button'
-export { SettingDropdown } from './dropdown/SettingDropdown'
-export { Header } from './header/Header'
-export { Logo } from './logo/Logo'
-export { Footer } from './footer/Footer'
+export {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from './dropdown/DropdownMenu'
 export { Heading } from './heading/Heading'
